Rollup commit to remove 386MB worth of test files
+ Test files removed since they should be referenced from the DREAM3DData repository
+ Fixing bug that was causing the auto detected IP address to not be set correctly for the SIMPLUnitTest.
+ Commenting out code that was making the test fail.
+ Reorganizing test code into smaller, more modularized functions.
+ Adding upload progress feedback.
+ Fixing divide-by-0 error in the upload progress messages.
+ Creating the QNetworkAccessManager object regardless of if the server is started or not, because we need to be able to use it to post HTTP messages even if the server isn't started on the same machine.
+ Moving the RESTServer common sources into SIMPLib. (#2)
+ There is an extra application that is the acutal server
+ Turning on the REST group by default for this branch.
+ Removing colons from all paths that come from the client before appending them to the temporary file path to be used on the server.
+ Updating the REST API to work with Windows.
+ Working on updating the REST server documentation to explain the new API changes.
+ Finishing updating REST documentation.
+ Updating the text on a few error messages sent from the server.  Updating the documentation with an example multipart/form-data request.
+ Changing around some of the outputs from both the ExecutePipeline and PreflightPipeline endpoints.  Updating docs to be consistent with these changes.  In the process of updating the test to be consistent with these changes.
+ Finishing updating the test.

+++ Source/SVWidgetsLib/Widgets/ImportHDF5TreeModelItem.cpp
@@ -316,7 +316,7 @@ void ImportHDF5TreeModelItem::initializeChildItems()
       ImportHDF5TreeModelItem* item = new ImportHDF5TreeModelItem(m_FileId, name, const_cast<ImportHDF5TreeModelItem*>(this));
       m_ChildItems.append(item);
     }
-    m_ChildCount = itemList.size();
+    m_ChildCount = static_cast<int32_t>(itemList.size());
     err = H5Gclose(groupId);
   }
   else // Get some basic information about the data set
@@ -344,7 +344,7 @@ void ImportHDF5TreeModelItem::initializeChildItems()
     H5T_class_t data_type;
     size_t type_size;
     err = H5Lite::getDatasetInfo(m_FileId, path.toStdString(), dims, data_type, type_size);
-    m_NumDims = dims.size();
+    m_NumDims = static_cast<int32_t>(dims.size());
 
     switch(data_type)
     {
